requirements.txt
certifi==2018.4.16
chardet==3.0.4
docker==3.3.0
docker-pycreds==0.2.3
grpcio==1.12.0
grpcio-tools==1.12.0
idna==2.6
protobuf==3.5.2.post1
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.48.0
